[BREAKING CHANGE] LookupInputWidget: Remove the silly 'input' parameter
Every usage I've found just passes 'this' there, and this code
wouldn't quite work properly if one passed something else than 'this'
(the parameters to TextInputMenuSelectWidget were passed incorrectly).

The parameter is confusing, making it seem like this class can be used
independently, while in fact it can only be used as a mixin to other
classes.

Bug: T85467
Change-Id: I33c818a2f18c8a3ce9e0d0cb89c2f6f9bf1f74d2