Seriously work around the Chromium scrollbar bug for good this time
When we remove 'overflow: auto' or 'overflow: scroll' from a DOM node,
Chromium wants to be so efficient recalculating the layout inside it
that it just doesn't do it at all. This is well-known and reported as
<https://code.google.com/p/chromium/issues/detail?id=387290>, and it
has a messy but working workaround.

But what's worse, it wants to be so efficient that it sometimes
doesn't recalculate it even when we do that workaround. I've
previously grapped with it in 141a0d7a (concerning MessageDialogs
being sized incorrectly because of this), it has now reappeared as
OptionWidgets not taking up the whole width of their parent
SelectWidgets.

Since this is so hopeless, let's extract and clean up the existing
better workaround from MessageDialog to a static method and use that
where needed (or possibly needed). Perhaps we should wrap it in a
browser check to avoid the performance penalty for less efficient, but
less broken browsers?

Change-Id: Id64c9972a13ccea2cde3f50fbb149779f4241587