TextInputWidget: Don't try adjusting size when detached
If we call setValue() on the TextInputWidget there before we attach
it, it calls adjustSize(), which calculates with bogus values and
caches them. Then we attach, adjustSize() is called again and does
nothing because cached values exist. This was probably introduced when
we added the caching thingy as performance improvement.

Let's track the attachedness state ourselves and only adjust size when
we're attached.

Bug: T85689
Change-Id: I6ca810c8d2e7968a5c6631b472e80c3a1dbf0c1e