MediaWiki theme: Fix background issues with disabled buttons
Followup to 04d04c0f.

The background was specified twice, once for the button itself and
once for the label. This caused funky-looking animation weirdness
because the label's background color wasn't animated but the button's
was.

However, we can't just remove the color setting on the label
because that breaks disabled frameless buttons, so this commit
also reorganizes the way color is set for disabled buttons to make
it more sane.

@disabled-text was changed from #fff to #eee and now only applies
to text in disabled frameless buttons. For disabled framed buttons,
we use the new @disabled-framed-text (as well as @disabled-background,
which is not applied to frameless buttons because those don't have
a background).

Bug: T87127
Change-Id: Ic0b76742b347fb976091811fbb1596eeed1fe47c