Add MenuLayout
Problem: GridLayout doesn't really solve the problem we need it to
solve, and is most likely an overengineered pile of JavaScript trash.

Solution: Add a new layout with two areas: the aside which can be
positioned on of the the sides and the content which fills the
remaining space. The menu can be toggled.

Change-Id: I86b9f6729d8eaf58509fb48c41835c5d8d7587a9