Remove usage of this.$ and config.$
OO.ui.Element's $ property is now deprecated. Until it's removed some
time in the future, the $ property, commonly known as "this dot dollar"
will be a reference to the global jQuery instance. Now that window
isolation is gone, we don't need to tediously inject a bound jQuery
function for every element.

Liberation from the oppressive this.$ regime!
Party in the streets! A round of waters for all my friends†.

Notes on how to get with the program:
* Replace use of this.$() with $()
* Remove $ properties from all config objects
* No longer use getElementDocument or $.context (you can just use
  document)

† Roan is buying

Change-Id: Iebe69e3617d7823240b86d49ad3b721641e7149e