MediaWiki theme: Consistent toggle button 'active' state
.oo-ui-buttonElement-active and .oo-ui-buttonElement-pressed is not
the same: the former is persistent, like in the toggle button; the
latter is temporary, as in a button currently being pressed by the user.
(Confusingly, CSS pseudoclass :active acts like our 'pressed' state.)

Apex styles them both identically, but in the MediaWiki theme we use
dark background for 'active' and light grey for 'pressed'. However,
this distinction was not implemented quite right everywhere.

Change-Id: I91df3da1bd83e83c4c52983cb2b02fe66583ac66