MediaWiki theme: Fix non-clickability of radios and checkboxes
The formatting <span/> was displayed on top of the (invisible) <input/>,
making it unclickable (unless there was a wrapping FieldLayout providing
a <label/>). Display the <input/> on top instead.

This also happens to fix an issue with hover state not working, as that
was caused by the :hover pseudoclass being styled on the <input/> rather
than the <span/> (which used to be incorrect and now will be correct).

Bug: T88035
Change-Id: Ide47e8eb05743aabde43a0a15633d09e2d2ea188