MenuSelectWidget: Codify current behavior of Tab closing the menu
Currently keyboard navigation of MenuSelectWidget only works when
used with a LookupElement/TextInputWidget. In that combination,
pressing Tab happens to close the menu, because it blurs the input,
which has a event handler to do the closing.

Codify this behavior in MenuSelectWidget itself too, since we need
to do something sensible with the menu when we make DropdownWidget
and such focusable, and the user opens the menu and tabs away. An
alternative to closing is selecting the highlighted option, which
is being discussed at T53229.

Change-Id: I21efc6f847dd09f44f0cb6833da7caca844928f7