DropdownWidget, ComboBoxWidget: Make keyboard-accessible
Make the dropdown box, or the little combo box arrow, focusable by
adding TabIndexedElement mixin. Handle both clicks and keypresses on
them.

The combo box arrow now does double duty, as an indicator for the text
input and as a tabindexed element for the combo box itself. This
appears to work without a hitch! We've built a really clever library.

Not included: focus state styles. Browser defaults will be used
(usually a blue outline).

Bug: T87705
Change-Id: I14b06ccf19a91d88b406d7e4e5bb92ef233ed421