PopupWidget: Set $clippable only once, correctly
Previously we didn't pass the $clippable option to the constructor,
setting the clippable element to this.$element (the ClippableElement
constructor called #setClippableElement). Then we built the DOM and
called #setClippableElement again ourselves, with this.$body, which
would then have to unset the existing clippable element first
(this.$element), which called OO.ui.Element.static.reconsiderScrollbars,
which is slow.

Bug: T89082
Change-Id: Ide21478276302bfb654d49563fec282bd3ee70be