Add "nofollow" option to ButtonWidget
Allow users to easily add rel="nofollow" to <a> hyperlinks to ensure that
the generated user interface isn't traversed by search engines.  This is
mostly relevant to the PHP variant of OOui, since the script-generated
elements generated by the JS variant won't typically be visible to search
engines in the first place.  But we add the configuration option to both
PHP and JS to keep them in sync.

The nofollow option defaults to true; the user must explicitly specify
nofollow=false if they want search engines to traverse their hyperlinks.

Change-Id: I57ab714b13e9c48fe82752922dcb1419bf7d9e9f