TextInputWidget: Deprecate 'icon' and 'indicator' events
The functionality they expose (user clicking on the icon/indicator) is
fundamentally not accessible: the icon/indicator is not focusable (has
no tabindex), has no keyboard events, doesn't have a label associated
or a tooltip, and has no sensible way of fixing all of this.

If something needs to be clickable separately, it should probably be a
separate Widget with the appropriate mixins added.

Example code change replacing the usage of 'indicator' event:
https://git.wikimedia.org/blobdiff/oojs%2Fui/9b327d95578cf7e76c704ad31e49727e7ae7c6d9/src%2Fwidgets%2FComboBoxWidget.js

Change-Id: I944bd1d3416b066b3283613d73fdf3dd61860c70