[BREAKING CHANGE] Make default distribution provide SVG with PNG fallback
Serves SVG images to modern browsers and PNG images as a fallback for
older ones, using the same technique we use for this in MediaWiki
(http://pauginer.tumblr.com/post/36614680636/invisible-gradient-technique).

This mixed SVG+PNG distribution is now the default one (without a
suffix in the name). The previously-default raster PNG-only distribution
is now available as with a '.raster' suffix, and the SVG-only distribution
is renamed to have a '.vector' suffix to match.

Bug: T87501
Change-Id: I5c56459fe6cf8690cc26eaac7969edd430251cab