MediaWiki Theme: Drop unnecessary pseudo-element of CheckboxInputWidget
The span's job was previously filled by pseudo-element and the
span was doing nothing but creating a bit of margin. (In a rather
strange way.) The margin keeps space between the widget and its label.
Apply this to the widget instead.

It was set to 0.2em on each side of the empty span (next to the
input). Together with the almost-1em of emptiness it was meant
to match the 5em margin-right that oo-ui-inline-spacing provides.

We should probably not be applying oo-ui-inline-spacing here, but,
for the moment, make this match 5em.

Change-Id: I5cc5e5acd2bffb1f56e76ed8a8849c352fa158bd