build: Use grunt-contrib-copy instead of custom 'copy' task
For our current build, it doesn't offer much. But this is in
prep for copying larger structures and edge cases (T90120) that
are much better implemented in this library.

Minor benefits:

* Performs slightly faster (YMMMV).

* Maintained by upstream.
  At least one minor bug is effectively fixed by this
  ("Copied 1 files > Copied 1 file")

* Match file descriptor syntax with that of other tasks (using
  standard Grunt file patterns instead of our own). This means
  they all work the same way  now (e.g. no more inconsistent
  handling of ** flattening or not, and trailing slash). This
  also addresses the weird behaviour where a copy would sometimes
  result in all files in the src directory being copied to a
  single target file taht was meant to be directory.

We are losing our built-in 'strip'. Created a rename function
in its place for now. I'd like to revise this setup more
generally, though. Our build can be simpler.

Change-Id: I49bc4be1dc564d195c9dbeec8ba4a6d8ca7780a5