MediaWiki theme: Simplify spacing of checkboxes/radios in FieldLayouts
Previously, a spacing of 1em between the widget and the label was
achieved by adding 0.5em margin-right on widget and 0.5em padding-left
on label. The 0.5em margin-right interferes with other uses, let's
zero it and use 1em padding-left on label instead. This is a lot
saner, too.

Also use the same method for RadioOptionWidget's label, which had a
smaller spacing previously. In Apex, too, for consistency; no visible
changes there.

Change-Id: I6bbb184a4268205cf788f86e817c390b0a2c1167