Move `OO.ui.infuse` to `OO.ui.Element.static.infuse`.
There is an "extra safe" typecheck in the code which is dependent on
T89721 being fixed in OOjs core and then released to OOjs UI.  It
is `if (false)`ed out with a FIXME for now, so that it doesn't
prevent the merge of this patch.  T90929 has been filed to
re-enable it once T89721 is complete.

Change-Id: I1999cbb2bdd1dd2557633882bf9fedc33788f6ce