Infer retry button action flags from symbolic name
Rather than storing an action object in this.currentAction, store the
symbolic action name, allowing executeAction to work even if triggered
programmatically rather than only when an actual button was clicked and
its reference stashed in this.currentAction.

If an error occurs, instead of assuming currentAction is a button, we
lookup buttons with the same action and copy flags form the first match
to the retry button.

Change-Id: I9d0c6c12c190431f3b0e3c874ce1909fd9c69900