FieldLayout: Fix display of bulleted list
The extra spaces at the start of the multi-line-items seem to be causing
the list to render incorrectly on the live site.

For some reason, the list looks fine locally; I'm not sure what is
different between the two setups that would make this the case. I'm also
not certain that removing the spaces will indeed fix the problem, but I
suspect it will.

Change-Id: I42eb1ad769b6128dcf2b35f13e13cb0ad2a4963c