build: Implement basic image flipping support in colorize-svg
I am not happy at all with this code, so much duplication. It
could probably be done a lot better if we rewrote the thing,
keeping in mind that we need this from the start. (Later today
in Things That Are Totally Gonna Happen, pigs flying.)

Follow-up to 06a5031fd3bf6f25f2398aff2b4999bfe224b463.

Change-Id: Id2e23894220015fffce27d8966d055e2f85824b2