[BREAKING CHANGE] MenuLayout: Reimplement without inline styles
Removes:
* 'menuSize' configuration option
* 'menuPositions' static property
* 'setMenuSize' and 'getMenuSize' methods
* 'updateSizes' method (private, but not documented as such)

The behavior of 'showMenu' and 'menuPosition' (wrongly documented as
'position') configuration options is not changed.

Menu size can be set using CSS now, using something along the lines of
the snippet below. MenuLayout's CSS will override the appropriate
values with 'auto' or '0' to display the menu correctly. If
`menuPosition` is known beforehand, CSS rules corresponding to other
positions may be omitted.

    .oo-ui-menuLayout-menu {
        height: 200px;
        width: 200px;
    }
    .oo-ui-menuLayout-content {
        top: 200px;
        left: 200px;
        right: 200px;
        bottom: 200px;
    }

Change-Id: I9ff531912505f897b75f39a6a8e2ffe435f6f829