Restore font family definitions to form elements
This reverts two thirds of commit 9aad519577, specifically the
removing of font rules from form elements, which don't inherit
from the body anyway.

In the case of form elements, use inherit for font-size and
font-family to achieve the desired effect of the original patch.

Also apply changes to Apex theme as well as MW.

Bug: T94450
Change-Id: I854c4bc134687838cd500cbe4c8a3f554a62085a