Revert "Button styles between OOJS and MW"
This reverts commit 626d1fe95f326e7b52dab7286baefe6578f2356d.

This caused no less than half a dozen different regressions.
I tried to fix it, but the changes done here are rather
invasive and disturbed the delicate balance of the nasty
styles we had here before, and it turned out to be hopeless.
Also, my FIXME comment was ignored, so I don't feel too bad.
Sorry.

Bug: T91523
Bug: T92553
Bug: T92556
Bug: T94199
Bug: T94201
Bug: T93472
Bug: T93002
Change-Id: I967f3abc64d44b9fe48073bb55e49755ba7663de