Introduce oo.ui.mixin namespace for mixins, and put them src/mixins
This helps disambiguate between (for example) ButtonWidget, which is an
instance of Widget and is instantiable, and GroupWidget which is a
non-instantiable mixin and not an instance of Widget.

Affected classes are *Element and {Group,Item}Widget.  The old
symbols have been retained as (deprecated) references to the new names.

Task: T94198
Change-Id: I7016697aaee4ad083d404c7b1b6963e0b4bad0fc