[DEPRECATING CHANGE] Create rtl-ready alignments in PopupWidget
Popup widget's alignment was defined as 'left', 'right' or
'center' which made it difficult to anticipate how these
will behave in RTL context. It was also unclear what 'left'
and 'right' refer to, as they seemed to refer to the placement
of the indicator rather than the leaninig of the popup itself,
which was confusing in both LTR and RTL.

For this reason, new alignments are created:

* force-left: The body of the popup is always leaning to
  the left, in both LTR and RTL.
* force-right: The body of the popup is always leaning to
  the right, in both LTR and RTL.
* backwards: The body of the popup is leaning towards the
  beginning of sentences in the given language direction.
* forwards: The body of the popup is leaning towards the
  end of the sentences in the given language direction.

The previously used terms 'right' and 'left' are deprecated
and are mapped to force-left and force-right, as their
context (alignment of the popup rather than alignment
of the indicator) has also changed.

Change-Id: Ia4570a36585995a2dcb0b329e9f052757f9b5316