[BREAKING CHANGE] Do not set font-size: 0.8em anywhere in the library
We were doing it for windows and some pieces of toolbars, but not
anything else. Using OOjs UI required you to counteract this if using
windows or toolbars. This has caused countless regressions and
immeasurable pain.

But this change is like a glitter bomb. We'll be finding pieces of
glitter in weird places for months.

OOjs UI is designed with font-size of 0.8em in mind (12.8px), and
that is the recommendation for users of the library, but it will work
correctly with any font-size (though icons may not be pixel-perfect).

Bug: T91152
Change-Id: Ib40e3477b764c3111205c232dc8958c5e2d2ff1b