Add IndexLayout
IndexLayout (new) is similar to BookletLayout but much simpler. It uses
CardLayout (new) instead of PageLayout, each of which relates to
generated TabOptionWidget (new) objects which are inside a new
TabSelectWidget (new) object.

The styling is gray background with white tabs that connect to the
content below. The unselected tabs have hover background color change.
The titles are bold like buttons. Tabs are left aligned and also an
active state was added to the tabs.

Bug: T97877
Change-Id: I3deaa165d1d18a83dff13f469caeda63f34e32b1