Switch over 'magnifyingGlass' icon to be 'search'
The 'search' icon was meant mostly as an indicator – it was tiny
and did not match other icons well. For historical reasons we had
both 'search' and 'magnifyingGlass' icons; replaced the latter with
the former for MediaWiki theme, and for both Apex and MediaWiki
themes made 'search' available as a proper indicator.

Bug: T97654
Change-Id: I6a001f388028d6041a7ab59d2f39daf9c15ea9c7