Simplify default action prevention in buttons and forms
I think this is beautifully simple: the default action of clicking a
button or submitting a form should be prevented only if there is an
event handler registered for it. EventEmitter#emit actually provides
the ability to figure that out, although we weren't using it anywhere.

This fixes all of the issues identified in T91132 and is unlikely to
cause problems in existing code.

FormLayout: We used to prevent the 'submit' event unconditionally,
  which meant that the underlying HTML <form> element couldn't
  actually be submitted.

ButtonInputWidget: We used to prevent the 'click' event unconditionally,
  which meant that the button couldn't be used to submit a form (since
  form 'submit' event is not fired if the 'click' event is prevented).

ButtonWidget: We used not to prevent the 'click' event when a 'href' was
  given, which was problematic for progressive enhancement ('href' had
  to be removed if the 'click' event was to be handled from JavaScript,
  which meant that the button could no longer be middle-clicked etc.).

Bug: T91132
Change-Id: Ia93621e7a68219ede1be7e6e47bac71c0c8f78b5