Rename icons: clear → cancel, closeInput → clear
closeInput → clear because 'closeInput' really does not make much sense,
it is meant to be the icon used for emptying (clearing) a text field.

clear → cancel because it conflicts now (and conflicts with MobileFrontend's
icons, and the icon is in fact being used for cancelling things (like next
to a progress bar to cancel the long-running action).

Bug: T97804
Change-Id: Idbd0da9fa19f942e9ab612b3c3472e1494643078