ClippableElement: Fix behavior of long clippables
Revert misguided bit of code from f5823566 which caused funny
behavior when the clippable was higher than the screen.

I am sincerely hoping that we will never need clipping with the top
edge of the screen (or: popups that pop up upwards), because it will
not work. (Conversely, clippables wider than the screen do not work
correctly now, but they can be clipped with either right or left
edge.)

If we ever decide we need that, then ClippableElement will need to
know to which edge the clippable is anchored to be able to
reasonably decide how to clip it.

Bug: T98506
Change-Id: I3b5925999e0ee5f46f3865abf17f184a5b7d0a55