ClippableElement: Fix behavior of clippables in nested scrollables
Apparently 'cOffset' already includes the 'scrollTop' when the
scrollable container is not the window, which is kind of reasonable in
a sick way, since it's supposed to be position relative to document,
and that position changes when the scrollable parent is scrolled.

The issue wasn't affecting VisualEditor because our nested scrollables
(dialogs) are always using fixed positioning and thus their position
doesn't change.

Bug: T98506
Change-Id: I756d1eafa8b9e0fdfa34906f76bac5998ebc6abc