SelectWidget: Listen to keypresses and jump to matching items
The algorithm for translating keypress events into a buffer to match
against items is:
* Escape, arrows, anything that opens or closes the list: clear
* Backspace: remove last character
* Anything with non-zero charCode:
  * If the buffer === the typed character, jump to the next item
    matching that character.
  * Otherwise append the character to the buffer. Then stay on the
    current item if it still matches, otherwise jump to the next
    matching item.
* 1.5 seconds after the last keypress, clear the buffer.

"Matching" an item is just fetching the label (.getLabel() if that
returns a string or $label.text() otherwise), then case-insensitive
match with "collapsed" whitespace. Both strings are NFC normalized if
the ES6 normalize() method is available.

Bug: T91126
Change-Id: I642a4621c714b84a9c14f0357d56d4acf306c21b