Avoid 'Potentially unsafe "href" attribute value' exceptions for relative URLs
parse_url() is technically not intended for relative URLs, and for
certain ones (such as '/wiki/%D9%85%D9%84%D8%AD%D9%82:1930') it
returns false (which, per documentation, means that it is considered
a "seriously malformed" URL).

Explicitly check for relative URLs beginning with a slash, rather than
expect parse_url to recognize them. This is probably still a bit too
paranoid but oh well.

Bug: T94900
Change-Id: If399fc40f21b33e9b8e72adbd7c160cec1b7c09d