DropdownInputWidget, RadioSelectInputWidget: Consistently call #cleanUpValue
By default, the primary purpose of #cleanUpValue is to cast everything
to strings.

On PHP side, we were calling #cleanUpValue when setting a value, but
not when building the list of possible values. This meant that, for
example, if all values were consistently passed in as integers, the
given 'value' would essentially be ignored.

On JS side, we weren't calling #cleanUpValue at all anywhere. This
meant that #getValue would return whatever was passed in to widget
config and not strings, contrary to documentation.

Reported by Florian when working on I47a86492.

Change-Id: I2d35669dadc355b443472c12b9e489a5accf4b43