TextInputWidget: Allow setting the HTML autocomplete attribute
In some cases (e.g. if the application provides an autocomplete list
by itself), it's needed to turn off the browser provided autocomplete
function with the autocomplete parameter. Until now, a config option
"autocomplete" is ignored.

Add a new config option autocomplete to the PHP and JS implementation of
TextInputWidget which defaults to false (no autocomplete attribute
will be added).

Bug: T104923
Change-Id: I6d7094bef106e8323051dd40a433c0e2d5f54444