tests: Twist the time in comparison tests in a different manner
The OO.ui.Element#updateThemeClasses is asynchronous and debounced
so that multiple state changes during widget initialization, which
would normally each require a theme classes update, require only one.

However, in comparison tests we need to wait for the updates to
complete, or better yet, we want them to be synchronous because that
makes life so much easier. Until now we ran the debounced method
manually for the root widget, but that obviously only works for
the root widget, and would cause the tests to fail in cryptic ways
when the root widget included other widgets which needed their theme
classes to be updated (I4e8e755e66916abffb90516519fe2b8fc4b48cd7,
Ie14a35fac70d62ff7d102caaa56654ebde11d7dd).

So let's try something else: make the OO.ui.Element#updateThemeClasses
method not be debounced or asynchronous for the tests. It's a bit more
evil, and requires a weird check in FlaggedElement (but we already have
such workarounds for methods being called before their class is done
constructing in various places in the library), but should be a bit
more reliable.

Change-Id: I4460919cd424a67fed98a06735b86e19f34b1fba