ButtonOptionWidget: Make it more difficult to set an inappropriate 'tabIndex'
ButtonOptionWidgets are not supposed to be focusable, since the parent
ButtonSelectWidget handles all keyboard input. Giving one a 'tabIndex'
(other than the default -1) breaks the ability to select options with
the keyboard.

Make it more difficult to do to prevent users from shooting themselves
in the leg by ignoring the 'tabIndex' passed in configuration options.
(It's still possible to call #setTabIndex after the initialization, if
someone really wants to.)

Change-Id: Iddaa1aa5b66384ef4f042ba7b7acfae3d04e25c2