HorizontalLayout: Introduce
HorizontalLayout arranges its contents in a single line (using
`display: inline-block` for its items), with small margins between
them. Convenient when you need to put a number of block-level widgets
(or layouts) on a single line next to each other.

We should probably replace spurious usage of .oo-ui-inline-spacing()
Less mixin for block-level widgets, as this is a better solution to
the problem of putting widgets next to each other, but that could be a
mildly breaking change for some uses.

Bug: T107868
Change-Id: I87a81b2c1cf084045c61517551abffb518925135