MenuSelectWidget: Call #updateItemVisibility in more cases
Binding to only 'keypress' misses some situations where the value
changes. A noticeable one is pressing Backspace (which doesn't seem to
fire 'keypress' at all, only 'keydown'), but also right click paste,
and so on.

Use the same list of events which we use in InputWidget, and which so
far has served us well. This is already debounced, so there's no worry
about multiple events per key press.

Change-Id: I73068664dd9be7f7aacbcdbdbed8c6b072935eb5