Theme: Add theme classes to $icon and $indicator only
Previously we added them to $element, which meant that they continued
to apply to nested widgets, which proved problematic in
PopupButtonWidget.

Also remove now-unnecessary CSS selectors.

This would be a breaking change for third-party themes, if any existed.

We need to call this.updateThemeClasses() in some more places, for the
benefit of tests (where updateThemeClasses is synchronous). If it is
not called after this.$icon (from IconElement) and this.$indicator are
set up, the theme classes on them will not be updated.

Bug: T109625
Change-Id: Ie93e4d6ed5637cdd95367729e1349a6da923f461