ClippableElement: Allow $clippableContainer to be different from $clippable
This separates the concepts of the container that is made scrollable
and shrunk when needed ($clippable) and the container whose height
is measured to ensure it remains visible ($clippableContainer).
By specifying $clippableContainer as the parent of $clippable,
it's possible to use fixed footers below the clipped area.

Renames of internal variables:
* this.$clippableContainer -> this.$clippableScrollableContainer
* this.onClippableContainerScrollHandler -> this.onClippableScrollHandler
* cc{Offset,Height,Width} -> sc{Offset,Height,Width}
* c{Offset,Width} -> cc{Offset,Width}

Change-Id: I711603de5c2b190c8e1c7e39504115a73bf42122