SelectFileWidget: Re-design to use a clearly clickable button
The drop target (showDropTarget) is now shown in addition to the rest
of the widget above it.

Drop target and select button are disabled if the widget is disabled
or file upload is not supported.

setActive and related code removed as it served no purpose.

Deprecation: renames dragDropUI to showDropTarget

Change-Id: Ieb66b56534dfe08df3307c7861dbe79daf30f7b4