Use 'ease' instead of 'ease-in-out' for CSS transitions
Timing 'ease' means both ease-in and ease-out, similar to ease-in-out.
However their implementation is slightly different.

The 'ease' timing is faster in the beginning than then end.
The "ease-in-out" timing on the other hand has the same curve for
both 'in' and 'out'.

Using 'ease' makes for a slightly more responsive experience.

Change-Id: I81f2208258057d1d958d28919866d7bb7133d665