BookletLayout, IndexLayout: Make autoFocus and focussing more reliable
Previously they almost never worked. (In particular, not if the
focusable was a OOUI button, and not if the layout wasn't outlined.)

Using OO.ui.findFocusable() introduced in
Ia49fe741e5d8a380ea60e6fb29286f4ebc77641d.

The #focusFirstFocusable appears to have always
been unused (and still is).

Bug: T107343
Change-Id: I4acbe694208b757bde0e647e9661476f9f14be6f