Update outline widget when current item is scrolled out of view
In continuous mode the stack layout now emits an event when the
current item scrolls out of view. By emitting an event users can
decide if they want to subsequently change currentItem.

In BookletLayout we listen to this event and update the outline
widget accordingly (which in turn updates currentItem in the stack
layout).

A flag is used to ensure the changing currentItem doesn't trigger
scrollElementIntoView while the user is still scrolling.

Item offsets are not cached between scroll events
as any number of DOM changes could affect the element heights.

Bug: T88385
Change-Id: I12906df5b311a9858ef6e8c6c79fc2bdaf23782b