[BREAKING CHANGE] Turn Element#gatherPreInfuseState into a static method
This is a prerequisite for being able to reuse parts of DOM when
infusing (T114408), which will improve interoperability with browsers'
form autofilling features and is likely to improve performance.

Because the DOM reuse would have to happen in the constructor, and
because it will destroy some of the state we're trying to preserve (in
particular, detaching a node will lose its focus), we need to do the
gathering before constructing the widget.

Change-Id: I429548772d7a829806b291825589960f2f32de0c