Refactor ComboBoxWidget into ComboBoxInputWidget
Changed ComboBoxWidget to inherit from TextInputWidget (rather than
embed one inside) and renamed to match. It's weird that we haven't
done this before, it simplifies a lot of code.

This is a backwards-compatible change. Old name has been kept as an
alias, all old methods and configs should work as before.

Updated references throughout the repository.

Change-Id: I3e38c331211ffafb72851e6a530075fdc75063cd