Remove small 0.1em vertical margin from buttons
This was written when OOjs UI was a wee lil' library and survived
several refactors. Today we have the various Layouts which manage
spacing, and this margin only makes it harder to reason about
dimensions of widgets and to line them up. It's almost invisible in
practice, except when it causes something to be offset by one pixel
and look nasty :(

Also removing several places where the margin was being removed to
make things line up.

Change-Id: I27888a7529effadc4706a09ee0df46e63d7d3b6d