build: Test PHP documentation with Doxygen via composer and make pass
This can be invoked with `composer run doc`. Have added a pre-doxygen filter
written by Timo Tijhof for MediaWiki which improves its reading of our code,
but it's imperfect even so. For example, Doxygen doesn't understand the type
`string[]` for @var (but does for @param) so I've modified those to work.

Change-Id: Ic64d3b46c1b9a6fe381d4c286325089620374bed