CapsuleMultiSelectWidget: Make the text field span all available area
The calculations only work correctly when the widget is attached to
DOM, which is basically impossible to detect (although we try in
TextInputWidget, with varying results).

To avoid problems here, we try really hard to re-run them when the
user interacts with the widget (which obviously means that it's
attached and visible).

Bug: T119781
Change-Id: I1a8d92cd853c7fd0ca113e5d5ce62eca9dc56d86