Implement OO.ui.alert() and OO.ui.confirm()
Two new convenience methods that pop up a MessageDialog with specified
message and appropriate buttons, named after the JavaScript methods
alert() and confirm() (but with a Promise-based interface).

Opening a MessageDialog to show the user a simple message takes a
whole lot of boilterplate code and a pile of promises; these methods
abstract this away. See documentation for details.

Also, tweak MessageDialog styles to better handle the case where no
'title' or no 'message' is given.

Bug: T117076
Change-Id: I5e61538d08de70bfce6c270c09ca04c7da028cc0