Really filter out unsafe URLs, but don't throw silly exceptions
Two changes:
* URLs such as 'java\nscript:alert()' are now correctly rejected.
* But they no longer cause an exception to be throw; rather, they are
  prefixed with './'.

This means that we now *really* are proofed against being given URLs
from user input (knock on wood), and we don't reject plain relative
URLs like 'index.php' (but just prefix them, which changes nothing).

I have rewritten quite a bit of the code to make it clearer and to
bring the PHP and JavaScript versions closer together.

Bug: T107623
Change-Id: I5972a16462c12ec470225e55c2e020179961d3b8