WindowManager: Only apply `top: 1em; bottom: 1em;` to non-fullscreen windows
This was already supposed to be overridden in WindowManager.less
(see `&-fullscreen > .oo-ui-dialog > .oo-ui-window-frame`), but that
rule has the same specificity as this one and it was earlier in the
file, so it was being ignored.

I think a lot of this CSS could be moved to WindowManager.less and
made to apply only to fullscreen or floating windows, but it would
have to be done carefully.

Bug: T122238
Change-Id: I50fb372cf3c9dcc03d3677ffd27f4ea37cdfe610