build: Only build one graphics distribution (mixed/vector/raster), not all
Honestly, there is no reason ever for anyone to use anything other
than the default 'mixed' distribution. But we can pretend for a bit
longer.

We now only build one of the CSS graphics distributions at a time.

* `grunt build` will build the 'mixed' one.
* `grunt build --graphics=vector` will build the 'vector' one.
* `grunt build --graphics=raster` will build the 'raster' one.

There are no more `.vector.css` and `.raster.css` suffixes, as there's
no more need for them.

As we can only build and use one graphics distribution at once, the
demo no longer has an option to choose one. It will just always use
the one that OOjs UI was built with.

Also, `grunt quick-build` is back to building the 'vector' distribution
(reverting 02eb23b4). Previously it would build the 'mixed' one with
no PNG images, because the `.vector.css` suffix was inconvenient. This
hack is no longer needed, as the suffix is gone.

Bug: T92464
Change-Id: Ic6309b4e98a9f855891dcd85b112d5326a50cf7a