ComboBoxInputWidget: Don't make the 'datalist' infusable
It makes no sense, it can't be infused by itself and it's not needed
for anything. We just need it to have an 'id' attribute. We probably
should have a more specific method than ensureInfusableId().

This is currently causing the PHP demo to throw silly errors when
trying out the "Infuse" feature.

Change-Id: I21585f205dd7380d0d45eb0339f5a1a5d6c930fa