Put '@keyframes' rules inside a mixin to avoid duplicating them
Both themes had the '@keyframes oo-ui-progressBarWidget-slide' rules
defined in widgets.less outside of any mixins, and since that file was
imported while generating pretty much any distribution CSS file, it
was duplicated in all of them, including icon packs and such.

It seems that Less understands @keyframes and knows that it has to
hoist the rule out of the nesting.

Change-Id: Ia50e1f98a5fc54f993144a5d36e0e89cb07d869d