styles: Use block rather than inline-block to avoid line height issues
Change some 'display: inline-block' to 'display: block' to prevent
unexpected height changes and gaps between dropdown handles and
dropdown lists. In all these cases there's no reason for 'inline-block'
(only the widget itself has to be, and not anything inside the widget).
The trick is that any 'inline-block' nodes are treated like characters
(like text) when laying out the page, which means they have 'line-height'
applied, and line-height values are somewhat unpredictable (they come
from the page that uses our widgets, and not our code). If the
'line-height' happens to be large, the widget's apparent height will
be bigger than what we set, and it's not obvious why this is the case.

Bug: T126911
Change-Id: If4cdfc172e82afd1986d3467a81ef6f2b2ded1ed