Bring some sanity to the styling of MenuSelectWidget inside different widgets
Remove code applying some weird inconsistent tweaks to the styling
of dropdown lists, which made every widget look slightly different,
and which only applied to widgets not using $overlay.

Try the "Dialog with dropdowns ($overlay test)" in the dialogs demo
to see what changes. Also increasing the height of that a bit, since
it looked unpleasantly cramped.

Change-Id: I9b4d963951a6b7278d9615ea9456205e85bcbee2