OO.ui.debounce: If an immediate timeout is already waiting, don't re-set it
OO.ui.debounce normally has to re-set the internal timeout every time
the function is called, so that the real call occurs only once at the
end, `wait` milliseconds after last debounced call.

But `wait` equal to 0 (an immediate timeout) is a special case. It
means that the real call will occur as soon as the work on current
thread finishes. Therefore, if an immediate timeout is already set,
but wasn't fired yet (indicated by `timeout` being non-null), we don't
need to do anything.

(The `immediate` parameter of the function is confusingly named and
not relevant here. It should be named `callAtStart` or something.)

Bug: T126786
Change-Id: Ia3a8fcda1bde76c9085bd8f6f5478dd43848e056